+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Fixed
 
 - Fixed an issue with `AutoScrollAction.scrollToItem` that caused the scrolled offset to be reset to the bottom when using `VerticalLayoutGravity.bottom`.
-- Fixed an issue with the `didPerform` closure of `AutoScrollAction.OnInsertedItem`, where the visible items were stale after adjusting the offset with no animation. 
+- Fixed issue with the `didPerform` closures of `AutoScrollAction.OnInsertedItem` and `AutoScrollAction.Pin`, where the visible items were stale after adjusting the offset with no animation. 
 
 ### Added