+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Fixed
 
-- Fixed an issue with `AutoScrollAction.scrollToItem` that caused the scrolled offset to be reset to the bottom when using `VerticalLayoutGravity.bottom`.
+- Fixed an issue with `AutoScrollAction.OnInsertedItem` and `AutoScrollAction.Pin` that caused the scrolled offset to be reset to the bottom when using `VerticalLayoutGravity.bottom`.
 - Fixed issue with the `didPerform` closures of `AutoScrollAction.OnInsertedItem` and `AutoScrollAction.Pin`, where the visible items were stale after adjusting the offset with no animation. 
 
 ### Added
@@ -15,7 +15,8 @@
 
 ### Internal
 
-- Adding a new demo to showcase mixing `AutoScrollAction.scrollToItem` with `VerticalLayoutGravity.bottom`.
+- Adding new demos to showcase mixing `AutoScrollAction.OnInsertedItem` and `AutoScrollAction.Pin` with `VerticalLayoutGravity.bottom`.
+- Adding a new `AutoScrollAction.Behavior` protocol.
 
 # Past Releases